+++ docs/changelog.md
@@ -4,6 +4,11 @@ All notable changes to Homeboy CLI are documented in this file.
 
 (This file is embedded into the CLI binary and is also viewable via `homeboy changelog`.)
 
+## Unreleased
+
+- Make release git.tag step idempotent to work with version bump tags
+- Add release pipeline hint after version bump tagging
+
 ## [0.30.7] - 2026-01-20
 
 ### Changed